Abfrage der Rechte bei DB Write Operationen

+++ pep/src/handle_post_requests/HandleDBWriteJurors.java
@@ -6,11 +6,13 @@ import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import data_management.Driver;
 
@@ -51,38 +53,60 @@ public class HandleDBWriteJurors extends HttpServlet {
 		}
 
 		Driver datenhaltung = new Driver();
+		HttpSession session = request.getSession();
+		String session_ID = (String)(session.getAttribute("session_id"));
+		if (session_ID != null)
+		{
+			try
+			{
+				String accountname_ID = datenhaltung.getSubCat("sessionmap", session_ID).get(0).get("accountname_ID");
+				String rolle = datenhaltung.getSubCat("account", accountname_ID).get(0).get("rollename_ID");
+				if (rolle.equals("Admin"))
+				{
+					if (push_into_db.containsKey("organisationseinheitname_ID")) {
+						String group = push_into_db.get("organisationseinheitname_ID");
+						push_into_db.remove("organisationseinheitname_ID");
+						datenhaltung.deleteRow("jurormap", "organisationseinheitname_ID", group);//
 
-		try {
-			if (push_into_db.containsKey("organisationseinheitname_ID")) {
-				String group = push_into_db.get("organisationseinheitname_ID");
-				push_into_db.remove("organisationseinheitname_ID");
-				datenhaltung.deleteRow("jurormap", "organisationseinheitname_ID", group);//
-
-				for (String juror : push_into_db.values()) {
-					String email[] = juror.split(":");
-					HashMap<String, String> jurormap = new HashMap<>();
-					jurormap.put("organisationseinheitname_ID", group);
-					jurormap.put("accountname_ID", email[1]);
-					datenhaltung.insertHashMap("jurormap", jurormap);
+						for (String juror : push_into_db.values()) {
+							String email[] = juror.split(":");
+							HashMap<String, String> jurormap = new HashMap<>();
+							jurormap.put("organisationseinheitname_ID", group);
+							jurormap.put("accountname_ID", email[1]);
+							datenhaltung.insertHashMap("jurormap", jurormap);
+						}
+					} else {
+						ArrayList<HashMap<String, String>> juroren = new ArrayList<>();
+						for (String juror : push_into_db.values()) {
+							String email[] = juror.split(":");
+							HashMap<String, String> juro = new HashMap<>();
+							juro.put("accountname_ID", email[1]);
+							juroren.add(juro);
+						}
+						datenhaltung.insertNewGroup(juroren);
+					}
+					
+					PrintWriter out = response.getWriter();
+					out.println("<script>");
+					out.println("window.open(\"/pep/home/show_groups\", \"_self\")");
+					out.println("</script>");
+					out.close();
 				}
-			} else {
-				ArrayList<HashMap<String, String>> juroren = new ArrayList<>();
-				for (String juror : push_into_db.values()) {
-					String email[] = juror.split(":");
-					HashMap<String, String> juro = new HashMap<>();
-					juro.put("accountname_ID", email[1]);
-					juroren.add(juro);
+				else
+				{
+					RequestDispatcher rd = request.getRequestDispatcher("/login");
+					rd.forward(request,  response);
 				}
-				datenhaltung.insertNewGroup(juroren);
 			}
-		} catch (SQLException e) {
-			e.printStackTrace();
+			catch (SQLException e)
+			{
+				e.printStackTrace();
+			}
 		}
-		PrintWriter out = response.getWriter();
-		out.println("<script>");
-		out.println("window.open(\"/pep/home/show_groups\", \"_self\")");
-		out.println("</script>");
-		out.close();
+		else
+		{
+			RequestDispatcher rd = request.getRequestDispatcher("/login");
+			rd.forward(request,  response);
+		}
 	}
-
 }
